Refactored the Query to Class concept into its own control instead of it occupying the main window.
