+++ SimpleClassCreator.Lib/Services/DtoGenerator.cs
@@ -96,10 +96,8 @@ namespace SimpleClassCreator.Lib.Services
 
             var cn = t.Name + "Dto";
 
-            var wcfOk = p.IncludeSerializeablePropertiesOnly && p.IncludeSerializableAttribute;
-
             //Class declaration
-            if (wcfOk) sb.AppendLine("[DataContract]");
+            //if (wcfOk) sb.AppendLine("[DataContract]");
 
             sb.Append("public class ").AppendLine(cn);
 
@@ -120,7 +118,7 @@ namespace SimpleClassCreator.Lib.Services
                 var tp = pi.PropertyType;
 
                 //Class properties
-                if (wcfOk) sb.AppendLine("[DataMember]");
+                //if (wcfOk) sb.AppendLine("[DataMember]");
 
                 sb.Append("public ").Append(GetTypeAsString(tp));
