+++ SimpleClassCreator.Ui/QueryToClassControl.xaml
@@ -1,0 +1,46 @@
+﻿<UserControl Name="QueryToClass"
+             x:Class="SimpleClassCreator.Ui.QueryToClassControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:SimpleClassCreator.Ui"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid Background="White" Margin="0,0,-55,-382">
+        <GroupBox Header="Step 1: Connection String" Height="100" HorizontalAlignment="Left" Margin="12,12,0,0" Name="groupBox1" VerticalAlignment="Top" Width="625">
+            <Grid>
+                <Button Content="Test" Height="65" HorizontalAlignment="Left" Margin="567,6,0,0" Name="btnConnectionStringTest" VerticalAlignment="Top" Width="40" Click="btnConnectionStringTest_Click" />
+                <ComboBox Height="65" HorizontalAlignment="Left" Margin="8,6,0,0" Name="cbConnectionString" VerticalAlignment="Top" Width="553" IsEditable="True" ItemsSource="{Binding}" DisplayMemberPath="ConnectionString" SelectedValuePath="ConnectionString" KeyDown="cbConnectionString_KeyDown" />
+            </Grid>
+        </GroupBox>
+        <GroupBox Header="Step 2: Class Source" Height="146" HorizontalAlignment="Left" Margin="12,118,0,0" Name="groupBox2" VerticalAlignment="Top" Width="625">
+            <Grid Height="123" Width="608">
+                <RadioButton Content="Query" Height="16" HorizontalAlignment="Left" Margin="161,6,0,0" Name="rbSourceTypeQuery" VerticalAlignment="Top" IsChecked="True" GroupName="rbgSourceType" />
+                <RadioButton Content="Table Name" Height="16" HorizontalAlignment="Left" Margin="78,6,0,0" Name="rbSourceTypeTableName" VerticalAlignment="Top" GroupName="rbgSourceType" IsChecked="True" />
+                <Label Content="Source type" Height="28" HorizontalAlignment="Left" Margin="6,1,0,0" Name="lblSourceType" VerticalAlignment="Top" Width="72" />
+                <TextBox Height="89" HorizontalAlignment="Left" Margin="6,28,0,0" Name="txtSource" VerticalAlignment="Top" Width="596" LostKeyboardFocus="txtSource_LostKeyboardFocus" AcceptsReturn="True" />
+            </Grid>
+        </GroupBox>
+        <GroupBox Header="Step 3: Options" Height="275" HorizontalAlignment="Left" Margin="12,270,0,0" Name="groupBox3" VerticalAlignment="Top" Width="625">
+            <Grid>
+                <CheckBox Content="Specify namespace" Height="16" HorizontalAlignment="Left" Margin="8,13,0,0" Name="cbIncludeNamespace" VerticalAlignment="Top" Width="127" Checked="cbIncludeNamespace_Checked" Unchecked="cbIncludeNamespace_Checked" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="135,10,0,0" Name="txtNamespace" VerticalAlignment="Top" Width="120" IsEnabled="False" />
+                <Label Content="Class name" Height="28" HorizontalAlignment="Left" Margin="12,162,0,0" Name="lblClassName" VerticalAlignment="Top" />
+                <TextBox Height="23" HorizontalAlignment="Left" Margin="83,165,0,0" Name="txtClassName" VerticalAlignment="Top" Width="120" TextChanged="txtClassName_TextChanged" MouseDown="txtClassName_MouseDown" />
+                <Button Content="Default" Height="23" HorizontalAlignment="Left" Margin="209,165,0,0" Name="btnClassNameDefault" VerticalAlignment="Top" Width="75" Click="btnClassNameDefault_Click" />
+                <Button Content="Default" Height="23" HorizontalAlignment="Left" Margin="260,10,0,0" Name="btnNamespaceDefault" VerticalAlignment="Top" Width="75" IsEnabled="False" Click="btnNamespaceDefault_Click" RenderTransformOrigin="-0.267,0.478" />
+            </Grid>
+        </GroupBox>
+        <Button Content="Generate Code" Height="39" HorizontalAlignment="Left" Margin="49,762,0,0" Name="btnGenerate" VerticalAlignment="Top" Width="186" Click="btnGenerate_Click" />
+        <CheckBox Content="Save as file on generation" Height="16" HorizontalAlignment="Left" Margin="246,712,0,0" Name="cbSaveFileOnGeneration" VerticalAlignment="Top" Checked="cbSaveFileOnGeneration_Checked" Unchecked="cbSaveFileOnGeneration_Checked" />
+        <TextBox Height="23" HorizontalAlignment="Left" Margin="325,778,0,0" Name="txtFileName" VerticalAlignment="Top" Width="120" IsEnabled="False" TextChanged="txtFileName_TextChanged" />
+        <CheckBox Content="Replace existing files" Height="16" HorizontalAlignment="Left" Margin="266,734,0,0" Name="cbReplaceExistingFiles" VerticalAlignment="Top" IsEnabled="False" />
+        <Label Content="File name" Height="28" HorizontalAlignment="Left" Margin="262,778,0,0" Name="lblFileName" VerticalAlignment="Top" IsEnabled="False" />
+        <TextBox Height="23" HorizontalAlignment="Left" Margin="325,752,0,0" Name="txtPath" VerticalAlignment="Top" Width="120" IsEnabled="False" />
+        <Label Content="Path" Height="28" HorizontalAlignment="Left" Margin="262,752,0,0" Name="lblPath" VerticalAlignment="Top" IsEnabled="False" />
+        <Button Content="Default" Height="23" HorizontalAlignment="Left" Margin="451,778,0,0" Name="btnFileNameDefault" VerticalAlignment="Top" Width="75" IsEnabled="False" Click="btnFileNameDefault_Click" />
+        <Button Content="Default" Height="23" HorizontalAlignment="Left" IsEnabled="False" Margin="451,751,0,0" Name="btnPathDefault" VerticalAlignment="Top" Width="75" Click="btnPathDefault_Click" />
+        <Button Content="Generate Columns" Height="38" HorizontalAlignment="Left" Margin="49,712,0,0" Name="btnGenerateGridViewColumns" VerticalAlignment="Top" Width="186" Click="btnGenerateGridViewColumns_Click" />
+    </Grid>
+</UserControl>
