+++ SimpleClassCreator.Ui/SimpleClassCreator.Ui.csproj
@@ -117,6 +117,9 @@
       <DependentUpon>DtoMakerControl.xaml</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
+    <Compile Include="QueryToClassControl.xaml.cs">
+      <DependentUpon>QueryToClassControl.xaml</DependentUpon>
+    </Compile>
     <Compile Include="ResultWindow.xaml.cs">
       <DependentUpon>ResultWindow.xaml</DependentUpon>
     </Compile>
@@ -157,6 +160,10 @@
   </ItemGroup>
   <ItemGroup>
     <Page Include="DtoMakerControl.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="QueryToClassControl.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
